Add files via upload
 Created a robust and efficient microservice that utilizes RESTful endpoints to 
perform CRUD (Create, Read, Update, Delete) operations on a specific resource, 
with 
such as user profiles. The project emphasizes the use of asynchronous 
programming 
Tokio 
for 
improved 
performance, 
data 
serialization/deserialization with Serde for effective data handling, and the 
implementation of a comprehensive error handling mechanism to ensure the 
reliability and resilience of the microservice. The primary goal is to build a scalable 
and maintainable microservice that provides seamless access to and management 
of the chosen resource through a RESTful API.